+++ chaincfg/params.go
@@ -254,7 +254,7 @@ type Params struct {
 	// Extension-related
 	Extension     ExtensionType
 	CommitteeSize uint32
-	Latency       uint32
+	Latency       float32
 	EpochSize     uint32
 }
